+++ app/src/main/java/com/ipvc/manut_smart/admin/Create_Tec.kt
@@ -1,16 +1,17 @@
-package com.ipvc.manut_smart
+package com.ipvc.manut_smart.admin
 
 import android.os.Bundle
 import androidx.activity.enableEdgeToEdge
 import androidx.appcompat.app.AppCompatActivity
 import androidx.core.view.ViewCompat
 import androidx.core.view.WindowInsetsCompat
+import com.ipvc.manut_smart.R
 
-class Create_Department : AppCompatActivity() {
+class Create_Tec : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         enableEdgeToEdge()
-        setContentView(R.layout.activity_create_department)
+        setContentView(R.layout.activity_create_tec)
         ViewCompat.setOnApplyWindowInsetsListener(findViewById(R.id.main)) { v, insets ->
             val systemBars = insets.getInsets(WindowInsetsCompat.Type.systemBars())
             v.setPadding(systemBars.left, systemBars.top, systemBars.right, systemBars.bottom)
