+++ app/src/main/java/com/ipvc/manut_smart/admin/SubMenuDepartmentAdminActivity.kt
@@ -1,20 +1,27 @@
-package com.ipvc.manut_smart
+package com.ipvc.manut_smart.admin
 
 import android.os.Bundle
+import android.widget.ImageView
 import androidx.activity.enableEdgeToEdge
 import androidx.appcompat.app.AppCompatActivity
 import androidx.core.view.ViewCompat
 import androidx.core.view.WindowInsetsCompat
+import com.ipvc.manut_smart.R
 
-class Register_device : AppCompatActivity() {
+class SubMenuDepartmentAdminActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         enableEdgeToEdge()
-        setContentView(R.layout.activity_register_device)
+        setContentView(R.layout.activity_sub_menu_department_admin)
         ViewCompat.setOnApplyWindowInsetsListener(findViewById(R.id.main)) { v, insets ->
             val systemBars = insets.getInsets(WindowInsetsCompat.Type.systemBars())
             v.setPadding(systemBars.left, systemBars.top, systemBars.right, systemBars.bottom)
             insets
+        }
+
+        val btnBack = findViewById<ImageView>(R.id.btnBack)
+        btnBack.setOnClickListener {
+            finish()
         }
     }
 }
