+++ app/src/main/java/com/ipvc/manut_smart/user/MenuUserActivity.kt
@@ -1,4 +1,4 @@
-package com.ipvc.manut_smart
+package com.ipvc.manut_smart.user
 
 import android.content.Intent
 import android.os.Bundle
@@ -8,12 +8,14 @@ import androidx.appcompat.app.AppCompatActivity
 import androidx.core.view.ViewCompat
 import androidx.core.view.WindowInsetsCompat
 import com.google.firebase.auth.FirebaseAuth
+import com.ipvc.manut_smart.LoginActivity
+import com.ipvc.manut_smart.R
 
-class MenuTechnicalActivity : AppCompatActivity() {
+class MenuUserActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         enableEdgeToEdge()
-        setContentView(R.layout.activity_menu_technical)
+        setContentView(R.layout.activity_menu_user)
 
         ViewCompat.setOnApplyWindowInsetsListener(findViewById(R.id.main)) { v, insets ->
             val systemBars = insets.getInsets(WindowInsetsCompat.Type.systemBars())
@@ -22,23 +24,28 @@ class MenuTechnicalActivity : AppCompatActivity() {
         }
 
         val user = FirebaseAuth.getInstance().currentUser
-
         if (user == null) {
+            // Sem sessão ativa — redireciona para login
             val intent = Intent(this, LoginActivity::class.java)
             intent.flags = Intent.FLAG_ACTIVITY_NEW_TASK or Intent.FLAG_ACTIVITY_CLEAR_TASK
             startActivity(intent)
             finish()
         }
 
+        // Referência ao botão de logout
         val logoutButton = findViewById<ImageView>(R.id.logout)
 
+        // Ação de logout
         logoutButton.setOnClickListener {
+            // Encerrar sessão do Firebase
             FirebaseAuth.getInstance().signOut()
 
+            // Redirecionar para a tela de login, limpando a pilha de Activities
             val intent = Intent(this, LoginActivity::class.java)
             intent.flags = Intent.FLAG_ACTIVITY_NEW_TASK or Intent.FLAG_ACTIVITY_CLEAR_TASK
             startActivity(intent)
 
+            // Finalizar a activity atual
             finish()
         }
     }
