+++ app/src/main/java/com/ipvc/manut_smart/user/Report_Avaria_Cli.kt
@@ -1,16 +1,17 @@
-package com.ipvc.manut_smart
+package com.ipvc.manut_smart.user
 
 import android.os.Bundle
 import androidx.activity.enableEdgeToEdge
 import androidx.appcompat.app.AppCompatActivity
 import androidx.core.view.ViewCompat
 import androidx.core.view.WindowInsetsCompat
+import com.ipvc.manut_smart.R
 
-class Devices_admin : AppCompatActivity() {
+class Report_Avaria_Cli : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         enableEdgeToEdge()
-        setContentView(R.layout.activity_devices_admin)
+        setContentView(R.layout.activity_report_avaria_cli)
         ViewCompat.setOnApplyWindowInsetsListener(findViewById(R.id.main)) { v, insets ->
             val systemBars = insets.getInsets(WindowInsetsCompat.Type.systemBars())
             v.setPadding(systemBars.left, systemBars.top, systemBars.right, systemBars.bottom)
