+++ app/src/main/res/layout/activity_finish_repair_technical.xml
@@ -6,7 +6,7 @@ android:id="@+id/main"
 android:layout_width="match_parent"
 android:layout_height="match_parent"
 android:fillViewport="true"
-tools:context=".Finish_repair_technical">
+tools:context=".technical.Finish_repair_technical">
 
 <LinearLayout
     android:layout_width="match_parent"
@@ -31,13 +31,13 @@ tools:context=".Finish_repair_technical">
             android:layout_height="100dp"
             android:layout_marginTop="16dp"
             android:src="@drawable/logo"
-            android:contentDescription="@string/logo" />
+            android:contentDescription="@string/Logo" />
 
         <TextView
             android:padding="20dp"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:text="@string/Finish_Repair"
+            android:text="@string/FinishRepair"
             android:textSize="20sp"
             android:textStyle="bold" />
     </LinearLayout>
