+++ app/src/main/res/layout/activity_list_repairs_cliente.xml
@@ -1,12 +1,11 @@
 <?xml version="1.0" encoding="utf-8"?>
 <ScrollView xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fillViewport="true"
-    tools:context=".List_repairs_cliente">
+    tools:context=".user.List_repairs_cliente">
 
     <LinearLayout
         android:layout_width="match_parent"
@@ -29,13 +28,13 @@
             android:layout_height="100dp"
             android:layout_marginTop="16dp"
             android:src="@drawable/logo"
-            android:contentDescription="@string/logo" />
+            android:contentDescription="@string/Logo" />
 
         <TextView
             android:padding="20dp"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:text="@string/repair_List"
+            android:text="@string/RepairList"
             android:textSize="20sp"
             android:textStyle="bold"/>
     </LinearLayout>
