+++ app/src/main/res/layout/activity_pending_repair.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fillViewport="true"
-    tools:context=".Pending_repair">
+    tools:context=".technical.Pending_repair">
 
     <LinearLayout
         android:layout_width="match_parent"
@@ -31,13 +31,13 @@
                 android:layout_height="100dp"
                 android:layout_marginTop="16dp"
                 android:src="@drawable/logo"
-                android:contentDescription="@string/logo" />
+                android:contentDescription="@string/Logo" />
 
             <TextView
                 android:padding="20dp"
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
-                android:text="@string/Pending_repairs"
+                android:text="@string/PendingRepairs"
                 android:textSize="20sp"
                 android:textStyle="bold" />
         </LinearLayout>
