+++ app/src/main/res/layout/activity_repair_history.xml
@@ -6,7 +6,7 @@ android:id="@+id/main"
 android:layout_width="match_parent"
 android:layout_height="match_parent"
 android:fillViewport="true"
-tools:context=".Repair_history">
+tools:context=".user.Repair_history">
 
 <LinearLayout
     android:layout_width="match_parent"
@@ -31,13 +31,13 @@ tools:context=".Repair_history">
             android:layout_height="100dp"
             android:layout_marginTop="16dp"
             android:src="@drawable/logo"
-            android:contentDescription="@string/logo" />
+            android:contentDescription="@string/Logo" />
 
         <TextView
             android:padding="20dp"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:text="@string/Repair_History"
+            android:text="@string/RepairHistory"
             android:textSize="20sp"
             android:textStyle="bold" />
     </LinearLayout>
