+++ app/src/main/res/layout/activity_report_avaria_cli.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="#FFFFFF"
-    tools:context=".Report_Avaria_Cli">
+    tools:context=".user.Report_Avaria_Cli">
 
     <LinearLayout
         android:layout_width="match_parent"
@@ -27,14 +27,14 @@
             android:layout_width="200dp"
             android:layout_height="100dp"
             android:layout_marginTop="16dp"
-            android:contentDescription="@string/logo"
+            android:contentDescription="@string/Logo"
             android:src="@drawable/logo" />
 
 
         <TextView
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:text="@string/Choose_device"
+            android:text="@string/ChooseDevice"
             android:textSize="18sp"
             android:textStyle="bold"
             android:textColor="#000000"
@@ -145,12 +145,12 @@
                     android:layout_width="48dp"
                     android:layout_height="48dp"
                     android:src="@drawable/smartwatch_icon"
-                    android:contentDescription="@string/smartwatch" />
+                    android:contentDescription="@string/Smartwatch" />
 
                 <TextView
                     android:layout_width="wrap_content"
                     android:layout_height="wrap_content"
-                    android:text="@string/smartwatch"
+                    android:text="@string/Smartwatch"
                     android:textColor="#FFFFFF"
                     android:layout_marginTop="8dp"/>
             </LinearLayout>
