+++ app/src/main/res/layout/activity_sub_menu_device_admin.xml
@@ -1,12 +1,11 @@
 <?xml version="1.0" encoding="utf-8"?>
 <ScrollView xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:fillViewport="true"
-    tools:context=".admin_technicians">
+    android:background="#FFFFFF"
+    tools:context=".admin.SubMenuDepartmentAdminActivity">
 
     <LinearLayout
         android:layout_width="match_parent"
@@ -15,69 +14,61 @@
         android:padding="24dp">
 
         <ImageView
+            android:id="@+id/btnBack"
             android:layout_width="40dp"
             android:layout_height="40dp"
             android:src="@drawable/return_icon"
             android:contentDescription="@string/Back" />
 
+
         <LinearLayout
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:orientation="vertical"
             android:gravity="center_horizontal">
+
 
             <ImageView
-                android:layout_width="200dp"
-                android:layout_height="100dp"
-                android:layout_marginTop="16dp"
-                android:src="@drawable/logo"
-                android:contentDescription="@string/logo" />
+                android:layout_width="350dp"
+                android:layout_height="150dp"
+                android:layout_marginTop="30dp"
+                android:contentDescription="@string/Logo"
+                android:src="@drawable/logo" />
 
             <TextView
-                android:padding="20dp"
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
-                android:text="Gestao de Tecnicos"
-                android:textSize="20sp"
-                android:textStyle="bold" />
-        </LinearLayout>
-
-        <LinearLayout
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:orientation="vertical"
-            android:padding="24dp"
-            android:gravity="center_horizontal">
+                android:layout_gravity="start"
+                android:layout_marginTop="25dp"
+                android:text="@string/DevicesManager"
+                android:textColor="@color/black"
+                android:textSize="25sp"
+                android:textStyle="bold"/>
 
             <Button
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
-                android:text="@string/Technicians_Register"
+                android:text="@string/AddDevice"
                 android:layout_marginTop="40dp"
                 style="@style/BlueButton"
                 android:height="50dp" />
-            <Button
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:text="@string/technicians_list"
-                android:layout_marginTop="10dp"
-                style="@style/WhiteButton"
-                android:height="50dp" />
-            <Button
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:text="@string/enable_technicians"
-                android:layout_marginTop="10dp"
-                style="@style/WhiteButton"
-                android:height="50dp" />
-            <Button
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:text="@string/technical_reports"
-                android:layout_marginTop="10dp"
-                style="@style/WhiteButton"
-                android:height="50dp" />
-        </LinearLayout>
 
+            <Button
+                style="@style/WhiteButton"
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                android:text="@string/DeviceList"
+                android:layout_marginTop="10dp"
+                android:height="50dp" />
+
+            <Button
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                android:text="@string/EnableDisableDevice"
+                android:layout_marginTop="10dp"
+                style="@style/WhiteButton"
+                android:height="50dp" />
+
+        </LinearLayout>
     </LinearLayout>
 </ScrollView>
