+++ app/src/main/res/values/strings.xml
@@ -1,79 +1,79 @@
 <resources>
     <string name="app_name">Manut_Smart</string>
-    <string name="title_activity_register">RegisterActivity</string>
+    <string name="TitleActivityRegister">RegisterActivity</string>
     <string name="Welcome">Bem-vindo</string>
-    <string name="email">Email</string>
+    <string name="Email">Email</string>
     <string name="Password">Palavra-Passe</string>
     <string name="Enter">Entrar</string>
     <string name="Register">Não tem conta? Registe-se</string>
-    <string name="Create_Account">Criar Conta</string>
+    <string name="CreateAccount">Criar Conta</string>
     <string name="Name">Nome</string>
-    <string name="Mobile_number">Número de Telemóvel</string>
-    <string name="Confirm_Password">Confirmar Palavra-Passe</string>
-    <string name="registarbtn">Registar</string>
-    <string name="Text_Register">Ja tem conta? Inicie sessao</string>
+    <string name="MobileNumber">Número de Telemóvel</string>
+    <string name="ConfirmPassword">Confirmar Palavra-Passe</string>
+    <string name="RegistarBtn">Registar</string>
+    <string name="TextRegister">Ja tem conta? Inicie sessao</string>
     <string name="Hi">Olá</string>
     <string name="IntroAdmin">Gestão de Avarias de Equipamentos</string>
     <string name="Repair">Reparações</string>
     <string name="Departement">Departamentos</string>
     <string name="Devices">Dispositivos</string>
     <string name="Tecnic">Técnicos</string>
-    <string name="logo">Logo</string>
-    <string name="logout">Logout</string>
-    <string name="Report_Fault">Reportar Avaria</string>
-    <string name="Ongoing_reports">Avarias em curso</string>
-    <string name="Fault_History">Histórico de Avarias</string>
+    <string name="Logo">Logo</string>
+    <string name="Logout">Logout</string>
+    <string name="ReportFault">Reportar Avaria</string>
+    <string name="OngoingReports">Avarias em curso</string>
+    <string name="FaultHistory">Histórico de Avarias</string>
     <string name="Back">Voltar</string>
-    <string name="Choose_device">Escolha o tipo de dispositivo com problema:</string>
+    <string name="ChooseDevice">Escolha o tipo de dispositivo com problema:</string>
     <string name="Computer">Computador</string>
     <string name="Mobile">Telemóvel</string>
     <string name="Tv">Televisão</string>
-    <string name="smartwatch">Smartwatch</string>
+    <string name="Smartwatch">Smartwatch</string>
     <string name="Console">Consola</string>
-    <string name="Urgency_of_service">Urgência do serviço:</string>
-    <string name="light">Leve</string>
-    <string name="average">Média</string>
-    <string name="high">Alta</string>
-    <string name="Devices_list">Lista de Dispositivos:</string>
+    <string name="UrgencyOfService">Urgência do serviço:</string>
+    <string name="Light">Leve</string>
+    <string name="Average">Média</string>
+    <string name="High">Alta</string>
+    <string name="DevicesList">Lista de Dispositivos:</string>
     <string name="Photo">Fotografias</string>
-    <string name="Problem_description">Descricao do problema:</string>
-    <string name="Load_Photo">Carregar fotografia</string>
+    <string name="ProblemDescription">Descricao do problema:</string>
+    <string name="LoadPhoto">Carregar fotografia</string>
     <string name="Cancel">Cancelar</string>
     <string name="Confirm">Confirmar</string>
-    <string name="Start_Repair">Iniciar Reparação</string>
-    <string name="Ongoing_repair">Reparação em curso</string>
-    <string name="Finish_repair">Finalizar Reparação</string>
-    <string name="repair_history">Histórico de Reparações</string>
-    <string name="plus">+</string>
-    <string name="repair_List">Lista de Avarias</string>
-    <string name="Pending_repairs">Reparações Pendentes</string>
+    <string name="StartRepair">Iniciar Reparação</string>
+    <string name="OngoingRepair">Reparação em curso</string>
+    <string name="FinishRepair">Finalizar Reparação</string>
+    <string name="RepairHistory">Histórico de Reparações</string>
+    <string name="Plus">+</string>
+    <string name="RepairList">Lista de Avarias</string>
+    <string name="PendingRepairs">Reparações Pendentes</string>
     <string name="Search">Pesquisar</string>
     <string name="Filter">Filtrar</string>
     <string name="Urgency">Urgência</string>
     <string name="Date">Data do pedido</string>
     <string name="Repair1">Reparar</string>
     <string name="Finish">Finalizar</string>
-    <string name="Finish_Repair">Finalizar Reparaçoes</string>
-    <string name="Repair_Description">Descrição da ReparaçãO</string>
+    <string name="RepairDescription">Descrição da ReparaçãO</string>
     <string name="Mesage">Mensagem ...</string>
-    <string name="Repair_History">Historico de Reparaçoes</string>
-    <string name="Create_Department">Criar Departamento</string>
-    <string name="Department_List">Lista de Departamentos</string>
-    <string name="Enable_Disable_Department">Ativar/ Desativar Departamento</string>
-    <string name="Device_Register">Registar Dispositivo</string>
-    <string name="Device_List">Lista de Dispositivos</string>
-    <string name="Enable_Disable_Device">Ativar / Desativar Dispositivo</string>
-    <string name="Devices_management">Gestao de Dispositivos:</string>
-    <string name="Technicians_Register">Registar Tecnico</string>
-    <string name="technicians_list">Lista de Tecnicos</string>
-    <string name="enable_technicians">Ativar / Desativar Tecnico</string>
-    <string name="technical_reports">Relatorio de Tecnico</string>
+    <string name="CreateDepartment">Criar Departamento</string>
+    <string name="DepartmentList">Lista de Departamentos</string>
+    <string name="EnableDisableDepartment">Ativar / Desativar Departamento</string>
+    <string name="DeviceRegister">Registar Dispositivo</string>
+    <string name="DeviceList">Lista de Dispositivos</string>
+    <string name="EnableDisableDevice">Ativar / Desativar Dispositivo</string>
+    <string name="DevicesManagement">Gestao de Dispositivos:</string>
+    <string name="TechniciansRegister">Registar Técnico</string>
+    <string name="TechniciansList">Lista de Técnicos</string>
+    <string name="EnableTechnicians">Ativar / Desativar Técnico</string>
+    <string name="TechnicalReports">Relatorio de Técnico</string>
     <string name="Description">Descrição</string>
     <string name="Loc">Localização</string>
     <string name="Number">Número de série</string>
-    <string name="Add_Device">Adicionar Dispositivo</string>
-    <string name="Devices_List">Lista de Disposittivos</string>
-    <string name="Tec_List">Lista de Tecnicos</string>
+    <string name="AddDevice">Adicionar Dispositivo</string>
+    <string name="TecList">Lista de Técnicos</string>
+    <string name="DepartmentManager">Gestão de Departamentos</string>
+    <string name="DevicesManager">Gestão de Dispositivos</string>
+    <string name="TechnicalManager">Gestão de Técnicos</string>"
 
     <style name="OngoingRepairs">Reparaçoes em curso</style>
 </resources>
